+++ R/globals.R
@@ -1,1 +1,1 @@
-utils::globalVariables(c("incl_nt", "neu_set"))
+utils::globalVariables(c("incl_nt", "neu_set", "rm_qm"))